appointment_id as BIGINT

+++ scripts/database/schema.mysql.sql
@@ -1713,7 +1713,7 @@ CREATE TABLE calendar_appointment (
 # ----------------------------------------------------------
 CREATE TABLE calendar_appointment_plugin (
     id INTEGER NOT NULL AUTO_INCREMENT,
-    appointment_id SMALLINT NOT NULL,
+    appointment_id BIGINT NOT NULL,
     plugin_key TEXT NOT NULL,
     config MEDIUMTEXT NULL,
     create_time DATETIME NOT NULL,
